Added ability to evaluate code intermittently throughout a training epoch.

+++ utils.py
@@ -145,7 +145,8 @@ def replace_fc_layer(config, model):
     model.head = nn.Linear(input_features, output_classes)
 
 
-def save_checkpoint(config, epoch, model, max_accuracy, optimizer, lr_scheduler, logger, validation_accuracy):
+def save_checkpoint(config, epoch, model, max_accuracy, optimizer, lr_scheduler, logger, validation_accuracy=None,
+                    eval_num=-1):
     save_state = {'model': model.state_dict(),
                   'optimizer': optimizer.state_dict(),
                   'lr_scheduler': lr_scheduler.state_dict(),
@@ -155,19 +156,24 @@ def save_checkpoint(config, epoch, model, max_accuracy, optimizer, lr_scheduler,
     if config.AMP_OPT_LEVEL != "O0":
         save_state['amp'] = amp.state_dict()
 
-    save_path = os.path.join(config.OUTPUT, f'ckpt_epoch_{epoch}.pth')
+    if eval_num == -1:
+        save_path = os.path.join(config.OUTPUT, f'ckpt_epoch_{epoch}.pth')
+    else:
+        save_path = os.path.join(os.path.join(config.OUTPUT, 'ckpt_epoch_%d_eval_%d.pth' % (epoch, eval_num)))
+
     logger.info(f"{save_path} saving......")
     torch.save(save_state, save_path)
     logger.info(f"{save_path} saved !!!")
 
     # Check if saved checkpoint is no longer in top-10
-    top_10_epochs = np.flip(np.argsort(validation_accuracy))[:10]
-    for file in os.listdir(config.OUTPUT):
-        if 'ckpt_epoch' in file:
-            epoch_num = file[11:file.find('.')]
-            epoch_num = int(epoch_num)
-            if epoch_num not in top_10_epochs:
-                os.remove(os.path.join(config.OUTPUT, file))
+    if validation_accuracy is not None:
+        top_10_epochs = np.flip(np.argsort(validation_accuracy))[:10]
+        for file in os.listdir(config.OUTPUT):
+            if 'ckpt_epoch' in file and 'eval' not in file:
+                epoch_num = file[11:file.find('.')]
+                epoch_num = int(epoch_num)
+                if epoch_num not in top_10_epochs:
+                    os.remove(os.path.join(config.OUTPUT, file))
